+++ Tabloid/client/src/components/Login/Register.js
