+++ Tabloid/client/src/components/Posts/PostForm.js
@@ -8,8 +8,8 @@ import {
     Input,
     Button,
 } from "reactstrap";
-import { PostContext } from "../providers/PostProvider";
-import { CategoryContext } from "../providers/CategoryProvider";
+import { PostContext } from "../../providers/PostProvider";
+import { CategoryContext } from "../../providers/CategoryProvider";
 import { useHistory } from "react-router-dom";
 
 export const PostForm = () => {
