+++ Tabloid/client/src/components/Posts/PostList.js
@@ -1,12 +1,13 @@
 import React, { useContext, useEffect } from "react";
-import { PostContext } from "../providers/PostProvider";
+import { PostContext } from "../../providers/PostProvider";
 import Post from "./Post";
 
-export const MyPostList = () => {
-    const { posts, getMyPosts } = useContext(PostContext);
+export const PostList = () => {
+    const { posts, getAllPosts } = useContext(PostContext);
 
     useEffect(() => {
-        getMyPosts();
+
+        getAllPosts();
     }, []);
 
     // useEffect dependency array with dependencies - will run if dependency changes (state)
@@ -20,12 +21,13 @@ export const MyPostList = () => {
     //     }
     // }, [searchTerms])
 
+
     return (
         <div className="container">
-
             <div className="row justify-content-center">
                 <div className="cards-column">
                     {posts.map((post) => {
+
                         return <Post key={post.id} post={post} />
                     })}
                 </div>
@@ -33,4 +35,4 @@ export const MyPostList = () => {
         </div>
     );
 };
-export default MyPostList;
+export default PostList;
